# frontend/docs/doc2md_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ingestify API",
    "description": "\nAPI assíncrona para conversão de documentos para Markdown usando Docling\n\n## Autenticação\n\nEsta API suporta dois métodos de autenticação:\n\n### 1. Bearer Token (JWT)\n1. Registre um usuário em `/auth/register`\n2. Faça login em `/auth/login` para obter o token\n3. Clique no botão **\"Authorize\"** (cadeado 🔒) no topo desta página\n4. Cole o token no campo \"bearerAuth\" (sem prefixo \"Bearer\")\n5. Clique em \"Authorize\"\n\n### 2. API Key\n1. Faça login para obter um token JWT\n2. Crie uma API Key em `/api-keys/`\n3. Clique no botão **\"Authorize\"** (cadeado 🔒)\n4. Cole a API Key no campo \"apiKeyAuth\"\n5. Clique em \"Authorize\"\n\nApós autorizar, todos os endpoints protegidos usarão automaticamente suas credenciais.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/admin/routing": {
      "get": {
        "tags": [
          "Admin - Engines"
        ],
        "summary": "Feature routes and their backlog",
        "operationId": "list_routes_admin_routing_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Routes Admin Routing Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/admin/routing/{feature}": {
      "put": {
        "tags": [
          "Admin - Engines"
        ],
        "summary": "Create or replace a feature's route",
        "operationId": "put_route_admin_routing__feature__put",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "feature",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Feature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Put Route Admin Routing  Feature  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Engines"
        ],
        "summary": "Remove a route (it drains back to today's path)",
        "operationId": "delete_route_admin_routing__feature__delete",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "feature",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Feature"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Route Admin Routing  Feature  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/engines/status": {
      "get": {
        "tags": [
          "Admin - Engines"
        ],
        "summary": "Dispatcher, in-flight work and backlog",
        "operationId": "engines_status_admin_engines_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Engines Status Admin Engines Status Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RouteStep": {
        "properties": {
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "engine_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Engine Ids"
          },
          "group_strategy": {
            "type": "string",
            "enum": [
              "priority",
              "fill_first"
            ],
            "title": "Group Strategy",
            "default": "priority"
          },
          "scale_out_after_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 86400.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Out After Seconds"
          },
          "when": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/When"
              },
              {
                "type": "null"
              }
            ]
          },
          "spend_cap": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpendCap"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "engine_ids"
        ],
        "title": "RouteStep"
      },
      "RouteUpdate": {
        "properties": {
          "steps": {
            "items": {
              "$ref": "#/components/schemas/RouteStep"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Steps"
          },
          "on_no_engine": {
            "type": "string",
            "enum": [
              "hold",
              "fail"
            ],
            "title": "On No Engine",
            "default": "hold"
          },
          "fail_after_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2592000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Fail After Seconds"
          },
          "max_attempts": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Attempts",
            "default": 3
          },
          "remote_allowed_for": {
            "type": "string",
            "enum": [
              "admins",
              "all"
            ],
            "title": "Remote Allowed For",
            "default": "admins"
          },
          "user_period_limit_usd": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,6}|(?=[\\d.]{1,13}0*$)\\d{0,6}\\.\\d{0,6}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Period Limit Usd"
          },
          "remote_data_notice": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Remote Data Notice"
          },
          "dispatcher_fallback": {
            "type": "string",
            "enum": [
              "local_direct",
              "hold"
            ],
            "title": "Dispatcher Fallback",
            "default": "local_direct"
          },
          "dispatcher_down_seconds": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 30.0,
            "title": "Dispatcher Down Seconds",
            "default": 120
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "current_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Password"
          }
        },
        "type": "object",
        "required": [
          "steps"
        ],
        "title": "RouteUpdate"
      },
      "SpendCap": {
        "properties": {
          "usd": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,6}|(?=[\\d.]{1,13}0*$)\\d{0,6}\\.\\d{0,6}0*$)"
              }
            ],
            "title": "Usd"
          },
          "window": {
            "type": "string",
            "enum": [
              "day",
              "period"
            ],
            "title": "Window"
          }
        },
        "type": "object",
        "required": [
          "usd",
          "window"
        ],
        "title": "SpendCap"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "When": {
        "properties": {
          "min_wait_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 604800.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Wait Seconds"
          },
          "min_backlog": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Backlog"
          }
        },
        "type": "object",
        "title": "When"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtido via /auth/login"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key criada via /api-keys/"
      }
    }
  }
}